Comments
Thoughts: I think you need to work out the logic of how two cards "Match." Is there some class that they have to have the same? Some attribute to the HTML? It may be hard if all the cards are hard coded and you have to account for every combination of cards the person could click. Instead, I would think of each card as an object. Maybe like class Card(id, isFlipped, startingClass), etc. Then, use event bubbling to capture the card they clicked. Store it in a variable if it's the first time they've clicked, and compare the variable if it's the second time they've clicked.